🧪 Fix CI by adding Node.js type reference to tests
Added `/// <reference types="node" />` to `frontend/src/utils/content.test.ts` to resolve TypeScript errors in CI when using the built-in Node.js test runner.
