+++ frontend/src/utils/parse-frontmatter.ts
@@ -1,0 +1,25 @@
+export function parseFrontmatter(raw: string): { meta: Record<string, string | string[]>; body: string } {
+  const match = raw.match(/^---\n([\s\S]*?)\n---\n([\s\S]*)$/);
+  if (!match) return { meta: {}, body: raw };
+
+  const meta: Record<string, string | string[]> = {};
+  for (const line of match[1].split("\n")) {
+    const idx = line.indexOf(":");
+    if (idx === -1) continue;
+    const key = line.slice(0, idx).trim();
+    let val = line.slice(idx + 1).trim();
+    // Remove surrounding quotes
+    if (val.startsWith('"') && val.endsWith('"')) {
+      val = val.slice(1, -1);
+    }
+    // Parse JSON arrays
+    if (val.startsWith("[")) {
+      try {
+        meta[key] = JSON.parse(val);
+        continue;
+      } catch { /* fall through */ }
+    }
+    meta[key] = val;
+  }
+  return { meta, body: match[2].trim() };
+}
